Fix bug in juce thread pool
This code should check if the thread pool has completed all jobs
while(pool.getNumJobs() > 0 );
Unfortunately this results in the following asm:

	00426BC8 mov 	eax,dword ptr [eax+20h]
	00426BCB nop	dword ptr [eax+eax]
	00426BD0 test	eax,eax
	00426BD2 jg	0426BD0

which only checks pool.getNumJobs() a single time, and then get locked
in a infinite loop between test and jg. This is because getNumJobs only
returns a single int variable representing the size of the array used by the
threadpool. As threads complete their jobs, the remove them from the
array and thus this variable is decreased, but on another thread. Thus
this variable needs to be declared volatile to force the optimizer to
recheck getNumJobs() every iteration of the loop.