Fix some bugs
The synth was unable to read a file because of what the working
directory was for the executable vs in visual studio. Added the file to
the resources folder to fix that.
There was a bug that the first LFO had a colour on top of its current
colour. Fixed that as well.